pyproject.toml
[tool.poetry]
name = "matflow-new"
version = "0.3.0a104"
description = "Computational workflows for materials science"
authors = ["aplowman <adam.plowman@manchester.ac.uk>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "matflow" }
]
exclude = ["matflow/data/demo_data"]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
pytest = {version = "^7.2.0", optional = true}
matplotlib = "^3.7"
hpcflow-new2 = "^0.2.0a145"

[tool.poetry.group.dev.dependencies]
pylint = "^2.12.2"
ipykernel = "^6.6.1"
pytest = "^7.2.0"
commitizen = "^2.20.3"
pre-commit = "^2.16.0"
black = "^22.3.0"
sphinx = ">=4,<7"
sphinx-click = "^4.4.0"
sphinx-jinja = "^2.0.1"
sphinx-copybutton = "^0.5.0"
sphinx-design = "^0.4.1"
sphinx-togglebutton = "^0.3.2"
pydata-sphinx-theme = "^0.13.3"

[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"
pexpect = "^4.8.0"
papermill = "^2.4.0"
ipykernel = "^6.25.1"

[tool.poetry.group.pyinstaller.dependencies]
pyinstaller = "^5.13.0"

[tool.poetry.extras]
test = ["pytest"]

[tool.poetry.scripts]
matflow = 'matflow.cli:cli'

[tool.poetry.plugins.pyinstaller40]
hook-dirs = "matflow.__pyinstaller:get_hook_dirs"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.3.0a104"
tag_format = "v$version"
version_files = [ 
    "pyproject.toml:version",
    "matflow/_version.py"
]
bump_message = "bump: $current_version → $new_version [skip ci]"

[tool.black]
line-length = 90

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
